フロントエンドの`page.tsx`にログアウトボタンを追加し、Supabaseのサインアウト機能を統合。ログインおよびサインアップページでのアクションを整理し、`actions.ts`ファイルを削除して新しいアクションファイルに移行。
